Classifier to suggest user to visit a place
We have trained a classifier to suggest the user that he/she should visit the place or not that is
visited by his/her Facebook friend based on the number of comments, likes, heart, hate, normal
reaction positivity in post caption. One more important feature we have considered here is if the
user has already visited the place or not. Unfortunately, there is no API currently available,
which can get the listed information from the user account. So have created dummy data, which
contains 10000 records. Date set contains the following columns.